+++ apps/cli/vitest.config.ts
@@ -11,15 +11,4 @@ export default defineConfig({
     testTimeout: 10000,
     mockReset: true,
   },
-  // resolve: {
-  //   alias: {
-  //     '@': resolve(__dirname, './src'),
-  //     '@/db': resolve(__dirname, '../../packages/utils/src/db'),
-  //     '@/logger': resolve(__dirname, '../../packages/utils/src/logger'),
-  //     '@/finance': resolve(__dirname, '../../packages/utils/src/finance'),
-  //     '@/schema': resolve(__dirname, '../../packages/utils/src/db/schema'),
-  //     '@ponti/utils': resolve(__dirname, '../../packages/utils/src'),
-  //     '@ponti/ai': resolve(__dirname, '../../packages/ai/src'),
-  //   },
-  // },
 })
